# repoquery: fix rich deps matching by using provide expansion from libdnf