This is meant to fix a bug where rich dependencies would match a package even if the require specified a different version range than doesn't match the version of the provide.

Uses new libdnf (PR linked below) that can expand package provides while filtering so that it doesn't haveto be implemented in dnf in a non-straightforward way that was causing it to match rich dependencies incorrectly. Further description is with the commits, along with some shortcomings this currently has.

https://github.com/rpm-software-management/libdnf/pull/762

Tests for this are here:
https://github.com/rpm-software-management/ci-dnf-stack/pull/577

https://bugzilla.redhat.com/show_bug.cgi?id=1534123
https://bugzilla.redhat.com/show_bug.cgi?id=1698034