@j-mracek 2019-11-08
@lukash Did you made any performance measurements?