@lukash 2019-11-08
@j-mracek I did not. From what I remember the from the bug was always quite slow - but I might remember that wrong. Are you concerned about anything in particular?