@j-mracek 2019-11-08
I would like to hear that there is no performance issue with the code change. The original code is slow, but it could be even worse.