@lukash 2019-12-10
I've done some rudimentary performance measurements. As a summary, it's mostly the same, the new version is slightly slower. There's a big performance hit for `--whatconflicts`, `--whatrecommends`, `--whatenhances`, `--whatsupplements` and `--whatsuggests`, because I've added NEVRA resolving for their arguments. Otherwise the new rich dependency resolving couldn't be used for them and they would have inconsistent behavior with the other switches.

Here's a table with the measurements. I've picked the `--whatrequires` and `--whatconflicts` options for the testing. The command run was `dnf repoquery --OPTION QUERY`. The options are in second to fourth column, 'old' is original dnf, 'new' is with the patches. Old `dnf repoquery --whatconflicts QUERY` was always around 2s, so that's where the major performance difference is. The added NEVRA resolving is causing it.

```
 QUERY     | old --whatrequires | new --whatrequires | new --whatconflicts |
-----------+--------------------+--------------------+---------------------+
 *server*  | 5m 5s              | 5m 14s             | 1m 25s              |
 *dnf*     | 1m 6s              | 1m 17s             | 24s                 |
 *libsolv* | 25s                | 30s                | 12s                 |
 libsolv   | 2s                 | 2s                 | 3s                  |
```