@lukash 2019-12-10
Oh, just to add: The stars in the QUERIES are there to increase the amount of matches for the queries, so that the times are longer and performance differences more significant. They're not the usual queries users would run most of the time, obviously.