@j-mracek 2019-12-13
@lukash Please could you rebase it?