@j-mracek 2019-12-13
I have discovered a performance problem with --whatconflicts:


time sudo dnf repoquery -C --disableplugin=* --whatconflicts $(rpm -qa --queryformat="%{name},") | 

The original dnf is able to perform the command in 3.835s but with patches in 18.262s.
Please could you provide improvements?