@lukash 2019-12-13
@j-mracek:
> I have discovered a performance problem with --whatconflicts

Please kindly read my https://github.com/rpm-software-management/dnf/pull/1432#issuecomment-563932852. Thanks.