@lukash 2020-01-14
> I believe that information in doc/api_queries.rst is outdated. @lukash Please could you take a look?

Could you be more specific?