@j-mracek 2020-01-14
> > I believe that information in doc/api_queries.rst is outdated. @lukash Please could you take a look?
> 
> Could you be more specific?

At lease please could you update a table between lines 90 and 115. For requires an argument type query/package sequence is not described.
I also believe that it is worth to mentioned that `requires=[pkg]` could provides a different result from `requires=pkg.provides + pkg.files`.