@lukash 2020-01-14
> At lease please could you update a table between lines 90 and 115. For requires an argument type query/package sequence is not described.

Ok, I see, I'll try to add that. I also see there's no mention of `conflicts` and the rich deps filters, I guess I should add those... But it's also going to get a bit verbose and repetitive given how the table is structured...

> I also believe that it is worth to mentioned that `requires=[pkg]` could provides a different result from `requires=pkg.provides + pkg.files`.

I suppose, but I'm not entirely sure how to phrase that and I'm also having a bit of a hard time remembering what exactly are the differences here (rich deps evaluation being the main difference I'm introducing), do you have any suggestions what to mention?