@lukash 2020-01-14
I've added a commit updating the documentation in `api_queries.rst`.